+++ tests/ecmascript/test-numconv-parse-explimit.js
@@ -7,8 +7,8 @@
 /*===
 1e1000000 number Infinity
 1e-1000000 number 0
-1e1000000000 number Infinity
-1e-1000000000 number 0
+1e10000000 number Infinity
+1e-10000000 number 0
 RangeError
 RangeError
 RangeError
@@ -46,13 +46,13 @@ function expLimitTest() {
     test('1000000');        // net value is 1e1000000 -> Infinity
     test('-1000000');       // net value is 1e-1000000 -> 0
 
-    // just at the limit
-    test('1000000000');        // Infinity
-    test('-1000000000');       // 0
+    // just at the current limit
+    test('10000000');
+    test('-10000000');
 
     // just over the current limit
-    test('1000000001');    // net value should be 1e1000000001 -> Infinity; InternalError instead
-    test('-1000000001');   // net value should be 1e-1000000001 -> 0; InternalError instead
+    test('10000001');    // net value should be 1e1000000001 -> Infinity; RangeError instead
+    test('-10000001');   // net value should be 1e-1000000001 -> 0; RangeError instead
 
     // over the current limit
     test('10000000000');