+++ RELEASES.rst
@@ -3229,7 +3229,9 @@ Planned
 
 * Fix 'defined but not used' warning for Windows (GH-1775)
 
-* Fix some Clang warnings (GH-1777)
+* Fix some Clang warnings by avoiding undefined behavior by default, define
+  DUK_USE_ALLOW_UNDEFINED_BEHAVIOR to reduce the explicit undefined behavior
+  checks for better footprint/performance (GH-1777, GH-1795)
 
 * Fix some compilation warnings triggered when DUK_NORETURN is not defined;
   the internal DUK_WO_NORETURN() is used to include unreachable dummy